Fix a warning:  format ‘%ld’ expects argument of type ‘long int’

+++ graf2d/graf/src/TCandle.cxx
@@ -284,7 +284,7 @@ int TCandle::ParseOption(char * opt) {
             if (isHorizontal) {fOption = (CandleOption)(fOption + kHorizontal);}
             strncpy(brOpen,"                ",brClose-brOpen+1); //Cleanup
 
-            sprintf(fOptionStr,"CANDLE%c(%ld)",direction,fOption);
+            sprintf(fOptionStr,"CANDLE%c(%ld)",direction,(long)fOption);
          } else {
             preset = 1;
             fOption = (CandleOption)(fOption + fallbackCandle);
@@ -342,7 +342,7 @@ int TCandle::ParseOption(char * opt) {
             if (isHorizontal) {fOption = (CandleOption)(fOption + kHorizontal);}
             strncpy(brOpen,"                ",brClose-brOpen+1); //Cleanup
 
-            sprintf(fOptionStr,"VIOLIN%c(%ld)",direction,fOption);
+            sprintf(fOptionStr,"VIOLIN%c(%ld)",direction,(long)fOption);
 
          } else {
             preset = 1;
